create compareServerVersion file
