+++ app/lib/rocketchat/methods/setUser.ts
@@ -3,7 +3,7 @@ import { InteractionManager } from 'react-native';
 import { setActiveUsers } from '../../../actions/activeUsers';
 import { setUser } from '../../../actions/login';
 import { store as reduxStore } from '../../auxStore';
-import { compareServerVersion } from '../../utils';
+import { compareServerVersion } from '../../methods/helpers/compareServerVersion';
 
 // TODO
 export function _setUser(this: any, ddpMessage: { fields: any; id: any; cleared: any }) {
