+++ app/lib/rocketchat/services/restApi.ts
@@ -16,9 +16,9 @@ import { TParams } from '../../../definitions/ILivechatEditView';
 import { store as reduxStore } from '../../auxStore';
 import { getDeviceToken } from '../../../notifications/push';
 import { getBundleId, isIOS } from '../../../utils/deviceInfo';
-import { compareServerVersion } from '../../utils';
 import roomTypeToApiType, { RoomTypes } from '../../methods/roomTypeToApiType';
 import sdk from './sdk';
+import { compareServerVersion } from '../../methods/helpers/compareServerVersion';
 
 export const createChannel = ({
 	name,
