+++ app/views/NewMessageView.tsx
@@ -16,9 +16,9 @@ import { IApplicationState, IBaseScreen, ISearch, TSubscriptionModel } from '../
 import I18n from '../i18n';
 import database from '../lib/database';
 import { CustomIcon } from '../lib/Icons';
+import { compareServerVersion } from '../lib/methods/helpers/compareServerVersion';
 import Navigation from '../lib/Navigation';
 import RocketChat from '../lib/rocketchat';
-import { compareServerVersion } from '../lib/utils';
 import UserItem from '../presentation/UserItem';
 import { withTheme } from '../theme';
 import { goRoom } from '../utils/goRoom';
