+++ app/views/RoomInfoEditView/index.tsx
@@ -22,7 +22,6 @@ import I18n from '../../i18n';
 import database from '../../lib/database';
 import { CustomIcon } from '../../lib/Icons';
 import RocketChat from '../../lib/rocketchat';
-import { compareServerVersion } from '../../lib/utils';
 import KeyboardView from '../../presentation/KeyboardView';
 import { TSupportedPermissions } from '../../reducers/permissions';
 import { ModalStackParamList } from '../../stacks/MasterDetailStack/types';
@@ -38,6 +37,7 @@ import { IAvatar } from '../../definitions/IProfileViewInterfaces';
 import sharedStyles from '../Styles';
 import styles from './styles';
 import SwitchContainer from './SwitchContainer';
+import { compareServerVersion } from '../../lib/methods/helpers/compareServerVersion';
 
 interface IRoomInfoEditViewState {
 	room: ISubscription;
