+++ app/views/RoomView/List/index.tsx
@@ -10,8 +10,8 @@ import { themes } from '../../../constants/colors';
 import ActivityIndicator from '../../../containers/ActivityIndicator';
 import { TAnyMessageModel, TMessageModel, TThreadMessageModel, TThreadModel } from '../../../definitions';
 import database from '../../../lib/database';
+import { compareServerVersion } from '../../../lib/methods/helpers/compareServerVersion';
 import RocketChat from '../../../lib/rocketchat';
-import { compareServerVersion } from '../../../lib/utils';
 import debounce from '../../../utils/debounce';
 import { animateNextTransition } from '../../../utils/layoutAnimation';
 import log from '../../../utils/log';
